refactor(util): refactor functions in util directory
백엔드의 util 디렉토리 내부 유틸 함수들을 구조화함.
