+++ back/util/auth/passport.js
@@ -1,8 +1,8 @@
 const passport = require('passport');
 const FacebookStrategy = require('passport-facebook').Strategy;
 const GoogleStrategy = require('passport-google-oauth').OAuth2Strategy;
-const { facebookConfig, googleConfig } = require('../config/passportConfig');
-const logInOrSignUp = require('../util/logInOrSignUp');
+const { facebookConfig, googleConfig } = require('../../config/passportConfig');
+const logInOrSignUp = require('./logInOrSignUp');
 
 passport.serializeUser((user, done) => {
   done(null, user);
@@ -13,10 +13,7 @@ passport.deserializeUser((user, done) => {
 });
 
 // 페이스북 로그인
-passport.use(
-  'facebookLogin',
-  new FacebookStrategy(facebookConfig, logInOrSignUp),
-);
+passport.use('facebookLogin', new FacebookStrategy(facebookConfig, logInOrSignUp));
 
 // 구글 로그인
 passport.use('googleLogin', new GoogleStrategy(googleConfig, logInOrSignUp));
