+++ back/util/auth/signUp.js
@@ -1,8 +1,8 @@
 const uuidv4 = require('uuid/v4');
-const { Users, Words, Op } = require('../db/models');
-const getRandomInt = require('../util/getRandomInt');
+const { Users, Words, Op } = require('../../db/models');
+const getRandomInt = require('../getRandomInt');
 
-async function signUp(user, profile) {
+async function signUp(profile) {
   // 사용되지 않는 단어들을 찾는다
   const words = await Words.findAll({
     where: {
@@ -12,15 +12,17 @@ async function signUp(user, profile) {
     },
   });
 
+  let newUser;
   // 사용 가능한 단어가 있는 경우
   if (words.length > 0) {
     const word = words[getRandomInt(0, words.length)];
-    user = await Users.create({
+    newUser = await Users.create({
       userId: profile.id,
       nickname: word.word,
     });
+
     Words.update(
-      { userId: user.id },
+      { userId: newUser.id },
       {
         where: {
           id: {
@@ -31,17 +33,17 @@ async function signUp(user, profile) {
     );
   } else {
     // 사용 가능한 단어가 없는 경우
-    user = await Users.create({
+    newUser = await Users.create({
       userId: profile.id,
       nickname: uuidv4(),
     });
     Words.create({
-      word: user.nickname,
-      userId: user.id,
+      word: newUser.nickname,
+      userId: newUser.id,
     });
   }
 
-  return user;
+  return newUser;
 }
 
 module.exports = signUp;
