--- back/util/getDataValues.js
@@ -1,5 +1,0 @@
-function getDataValues(result) {
-  return result.dataValues;
-}
-
-module.exports = getDataValues;
